+++ lib/alf/tools/tuple_scope.rb
@@ -1,20 +1,21 @@
 module Alf
   module Tools
     #
-    # Provides a handle for implementing a flyweight design pattern on tuples.
+    # Specialization of Scope for working on a tuple specifically.
     #
-    # This class works hand-in-hand with TupleExpression as it provides a scope
-    # for their evaluation on tuples.
+    # This class works hand-in-hand with TupleExpression as it provides a scope for their
+    # evaluation on tuples. It also provides a handle for implementing a flyweight design
+    # pattern on tuples through the `__set_tuple` method.
     #
     # Example:
     #
-    #   handle = TupleHandle.new
-    #   expr   = TupleExpression["status > 10"]
+    #   scope = TupleScope.new
+    #   expr  = TupleExpression["status > 10"]
     #   relation.each do |tuple|
-    #     expr.evaluate(handle.set(tuple))
+    #     expr.evaluate(scope.__set_tuple(tuple))
     #   end
     #
-    class TupleHandle < Scope
+    class TupleScope < Scope
 
       module OwnMethods
 
@@ -26,20 +27,21 @@ module Alf
 
         # Sets the next tuple to use.
         #
-        # On first call, this method installs the handle as a side effect.
+        # On first call, this method installs the scope as a side effect unless a tuple
+        # has already been provided at construction.
         #
         # @param [Tuple] tuple the current iterated tuple
-        # @return [TupleHandle] self
+        # @return [TupleScope] self
         def __set_tuple(tuple)
           __build(tuple) if @tuple.nil?
           @tuple = tuple
           self
         end
 
-        # Builds this handle with a tuple.
+        # Builds this scope with a tuple.
         #
         # This method should be called only once and installs instance methods on
-        # the handle with keys of _tuple_.
+        # the scope with keys of _tuple_.
         def __build(tuple)
           tuple.keys.each do |k|
             (class << self; self; end).send(:define_method, k) do
@@ -51,12 +53,12 @@ module Alf
 
       end # module OwnMethods
 
-      # Creates a handle instance
+      # Creates a scope instance
       def initialize(tuple = nil, extensions = [], parent = nil)
         super [ OwnMethods ] + extensions, parent
         __build(@tuple = tuple) if tuple
       end
 
-    end # class TupleHandle
+    end # class TupleScope
   end # module Tools
 end # module Alf