+++ spec/unit/alf-core/types/test_summarization.rb
@@ -84,11 +84,11 @@ module Alf
     end # describe 
 
     specify "least -> happens -> finalize" do
-      handle = Tools::TupleHandle.new
+      scope = Tools::TupleScope.new
       summ = Summarization.coerce(["s", "sum{ qty }", "m", "max{ size }"])
       (x = summ.least).should eql(:s => 0, :m => nil)
-      (x = summ.happens(x, handle.__set_tuple(:qty => 10, :size => 12))).should eq(:s => 10, :m => 12)
-      (x = summ.happens(x, handle.__set_tuple(:qty => 5, :size => 5))).should eq(:s => 15, :m => 12)
+      (x = summ.happens(x, scope.__set_tuple(:qty => 10, :size => 12))).should eq(:s => 10, :m => 12)
+      (x = summ.happens(x, scope.__set_tuple(:qty => 5, :size => 5))).should eq(:s => 15, :m => 12)
       summ.finalize(x).should eq(:s => 15, :m => 12)
     end